Create sendEmail module. Use sendEmail module for password recovery.
