+++ views/passwordRecovery.hbs
@@ -21,12 +21,16 @@
       alertBox.innerHTML = "Please input your email address above";
       alertBox.style.display = "block";
     } else {
+      alertBox.className = "alert alert-info";
+      alertBox.innerHTML = "Email sending. Please wait.";
+      alertBox.style.display = "block";
       var req = new XMLHttpRequest();
       req.open('PUT', '/password_recovery', true);
       req.setRequestHeader('Content-Type', 'application/json');
 
       var formData = JSON.stringify({email});
       req.send(formData);
+
 
       req.addEventListener('load', function() {
         if (req.status == 200) {
